+++ lib/monolens/file.rb
@@ -2,13 +2,8 @@ module Monolens
   class File
     include Lens
 
-    def initialize(info)
-      super
-      options[:lenses] = Monolens.lens(options[:lenses])
-    end
-
     def call(*args, &bl)
-      options[:lenses].call(*args, &bl)
+      option(:lenses, ->(arg){}).call(*args, &bl)
     end
   end
 end