+++ lib/monolens/object/rename.rb
@@ -7,7 +7,7 @@ module Monolens
         is_hash!(arg)
 
         dup = arg.dup
-        options.each_pair do |oldname, newname|
+        option(:defn).each_pair do |oldname, newname|
           actual_name, value = fetch_on(oldname, arg)
           newname = actual_name.is_a?(Symbol) ? newname.to_sym : newname.to_s
           dup.delete(actual_name)