+++ lib/monolens/object/transform.rb
@@ -3,18 +3,19 @@ module Monolens
     class Transform
       include Lens
 
-      def initialize(transformation)
-        super({})
-        @transformation = transformation.each_with_object({}){|(attr,lens),memo|
-          memo[attr.to_sym] = Monolens.lens(lens)
-        }
+      def initialize(options)
+        super(options)
+        ts = option(:defn, {})
+        ts.each_pair do |k,v|
+          ts[k] = Monolens.lens(v)
+        end
       end
 
       def call(arg, *rest)
         is_hash!(arg)
 
         dup = arg.dup
-        @transformation.each_pair do |attr, sub_lens|
+        option(:defn, {}).each_pair do |attr, sub_lens|
           actual_attr, fetched = fetch_on(attr, dup)
           dup[actual_attr] = sub_lens.call(fetched)
         end