+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+## 0.3.0
+
+* BREAKING: object.select, object.rename and object.transform
+  now expect their selection/renaming/transformation to be
+  provided under a `defn` key. This allows for adding options
+  later.
+
+  Before:
+
+  ```yaml
+  - object.select:
+      firstname: name
+  ```
+
+  After
+
+  ```yaml
+  - object.select
+      defn:
+        firstname: name
+  ```
+
 ## 0.2.0 - 2022-05-05
 
 * BREAKING: (core.)map moved as array.map, where it belongs